# apbf: Switch to fast reduce method.