--- a/container/apbf/filter.go
+++ b/container/apbf/filter.go
@@ -10,6 +10,7 @@ package apbf
 import (
 	"encoding/binary"
 	"math"
+	"math/bits"
 	"sync"
 	"time"
 
@@ -302,6 +303,34 @@ func (f *Filter) setBit(bit uint64) {
 	f.data[bit>>3] |= 1 << (bit & 7)
 }
 
+// fastReduce calculates a mapping that is more or less equivalent to x mod N.
+// However, instead of using a mod operation that can lead to slowness on many
+// processors when not using a power of two due to unnecessary division, this
+// uses a "multiply-and-shift" trick that eliminates all divisions as described
+// in a blog post by Daniel Lemire, located at the following site at the time
+// of this writing:
+// https://lemire.me/blog/2016/06/27/a-fast-alternative-to-the-modulo-reduction/
+//
+// Since that link might disappear, the general idea is to multiply by N and
+// shift right by log2(N).  Since N is a 64-bit integer in this case, it
+// becomes:
+//
+// (x * N) / 2^64 == (x * N) >> 64
+//
+// This is a fair map since it maps integers in the range [0,2^64) to multiples
+// of N in [0, N*2^64) and then divides by 2^64 to map all multiples of N in
+// [0,2^64) to 0, all multiples of N in [2^64, 2*2^64) to 1, etc.  This results
+// in either ceil(2^64/N) or floor(2^64/N) multiples of N.
+func fastReduce(x, N uint64) uint64 {
+	// This uses math/bits to perform the 128-bit multiplication as the compiler
+	// will replace it with the relevant intrinsic on most architectures.
+	//
+	// The high 64 bits in a 128-bit product is the same as shifting the entire
+	// product right by 64 bits.
+	hi, _ := bits.Mul64(x, N)
+	return hi
+}
+
 // Add inserts the provided data into the filter.
 //
 // This function is safe for concurrent access.
@@ -335,7 +364,7 @@ func (f *Filter) Add(data []byte) {
 	hash1, hash2 := siphash.Hash128(f.key0, f.key1, data)
 	derivedIdx, acc := deriveIndex(logicalSlice, hash1, hash2)
 	for i := uint8(0); i < f.k; i++ {
-		f.setBit(sliceBitOffset + derivedIdx%f.bitsPerSlice)
+		f.setBit(sliceBitOffset + fastReduce(derivedIdx, f.bitsPerSlice))
 
 		// Move to the next logical slice while wrapping around the ring buffer
 		// if needed.
@@ -404,7 +433,7 @@ func (f *Filter) Contains(data []byte) bool {
 	hash1, hash2 := siphash.Hash128(f.key0, f.key1, data)
 	derivedIdx, acc := deriveIndex(logicalSlice, hash1, hash2)
 	for {
-		if f.isBitSet(sliceBitOffset + derivedIdx%f.bitsPerSlice) {
+		if f.isBitSet(sliceBitOffset + fastReduce(derivedIdx, f.bitsPerSlice)) {
 			// Successful query when the required number of consecutive matches
 			// is achieved.
 			curMatches++