--- a/container/apbf/go.mod
+++ b/container/apbf/go.mod
@@ -1,5 +1,5 @@
 module github.com/decred/dcrd/container/apbf
 
-go 1.11
+go 1.13
 
 require github.com/dchest/siphash v1.2.2